+++ Helper/Data.php
@@ -49,14 +49,12 @@ class Data extends AbstractHelper
      */
     protected $productMetadata;
 
-
     public function __construct(
         Context $context,
         \HiPay\FullserviceMagento\Model\RuleFactory $ruleFactory,
         ResourceInterface $moduleResource,
         ProductMetadataInterface $productMetadata
-    )
-    {
+    ) {
         parent::__construct($context);
         $this->ruleFactory = $ruleFactory;
         $this->moduleResource = $moduleResource;
@@ -146,4 +144,36 @@ class Data extends AbstractHelper
         return json_encode($request);
     }
 
+    /**
+     * @return bool
+     */
+    public function useOrderCurrency()
+    {
+
+        return (bool)$this->scopeConfig->getValue(
+            'hipay/configurations/currency_transaction',
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE,
+            null
+        );
+
+    }
+
+    /**
+     * @param \HiPay\FullserviceMagento\Model\Config $config
+     * @param \HiPay\FullserviceMagento\Model\Gateway\Manager $gatewayClient
+     * @param $store
+     * @param string $scope
+     * @return mixed
+     */
+    public function updateHashAlgorithm(
+        \HiPay\FullserviceMagento\Model\Config $config,
+        \HiPay\FullserviceMagento\Model\Gateway\Manager $gatewayClient,
+        $store,
+        $scope = \Magento\Store\Model\ScopeInterface::SCOPE_STORES
+    ) {
+        $hash = $gatewayClient->requestSecuritySettings();
+        $config->setHashingAlgorithm($hash, $scope);
+        $store->resetConfig();
+        return $hash;
+    }
 }