+++ Model/Config/AbstractConfig.php
@@ -67,15 +67,24 @@ abstract class AbstractConfig implements ConfigInterface
      */
     protected $_scopeConfig;
 
+    /**
+     * Core config writer
+     *
+     * @var \Magento\Framework\App\Config\Storage\WriterInterface
+     */
+    protected $_configWriter;
 
     /**
+     * AbstractConfig constructor.
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
+     * @param \Magento\Framework\App\Config\Storage\WriterInterface $configWriter
      */
     public function __construct(
-        \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
-    )
-    {
+        \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
+        \Magento\Framework\App\Config\Storage\WriterInterface $configWriter
+    ) {
         $this->_scopeConfig = $scopeConfig;
+        $this->_configWriter = $configWriter;
     }
 
     /**
@@ -179,6 +188,20 @@ abstract class AbstractConfig implements ConfigInterface
         );
     }
 
+    public function setGeneralValue(
+        $key,
+        $data,
+        $group = 'hipay_credentials',
+        $scope = \Magento\Store\Model\ScopeInterface::SCOPE_STORES
+    ) {
+        $this->_configWriter->save(
+            $this->_mapGeneralFieldset($key, $group),
+            $data,
+            $scope,
+            $this->_storeId
+        );
+    }
+
     /**
      * Store ID setter
      *
@@ -235,7 +258,11 @@ abstract class AbstractConfig implements ConfigInterface
             case 'secret_passphrase_test':
             case 'fingerprint_enabled':
             case 'basket_enabled':
+            case 'send_notification_url':
             case 'basket_attribute_ean':
+            case 'currency_transaction':
+            case 'hashing_algorithm':
+            case 'hashing_algorithm_test':
                 return "hipay/{$group}/{$fieldName}";
             default:
                 return null;