+++ Model/Request/Order.php
@@ -115,10 +115,23 @@ class Order extends CommonRequest
         \Magento\Customer\Api\CustomerRepositoryInterface $customerRepositoryInterface,
         \Magento\Customer\Api\GroupRepositoryInterface $groupRepositoryInterface,
         $params = []
-    )
-    {
-        parent::__construct($logger, $checkoutData, $customerSession, $checkoutSession, $localeResolver, $requestFactory,
-            $urlBuilder, $helper, $cartFactory, $weeeHelper, $productRepositoryInterface, $mappingCategoriesCollectionFactory, $categoryFactory, $params);
+    ) {
+        parent::__construct(
+            $logger,
+            $checkoutData,
+            $customerSession,
+            $checkoutSession,
+            $localeResolver,
+            $requestFactory,
+            $urlBuilder,
+            $helper,
+            $cartFactory,
+            $weeeHelper,
+            $productRepositoryInterface,
+            $mappingCategoriesCollectionFactory,
+            $categoryFactory,
+            $params
+        );
 
         $this->helper = $helper;
         $this->_cartFactory = $cartFactory;
@@ -202,6 +215,7 @@ class Order extends CommonRequest
     protected function mapRequest()
     {
         $payment_product = $this->getSpecifiedPaymentProduct();
+        $useOrderCurrency = $this->_config->useOrderCurrency();
 
         $orderRequest = new OrderRequest();
         $orderRequest->orderid = $this->_order->getForcedOrderId() ?: $this->_order->getIncrementId();
@@ -209,10 +223,18 @@ class Order extends CommonRequest
         $orderRequest->payment_product = $this->getCcTypeHipay($this->_order->getPayment()->getCcType()) ?: $payment_product;
         $orderRequest->description = $this->_order->getForcedDescription() ?: sprintf("Order %s", $this->_order->getIncrementId()); //@TODO
         $orderRequest->long_description = "";
-        $orderRequest->currency = $this->_order->getBaseCurrencyCode();
-        $orderRequest->amount = $this->_order->getForcedAmount() ?: (float)$this->_order->getBaseGrandTotal();
-        $orderRequest->shipping = (float)$this->_order->getShippingAmount();
-        $orderRequest->tax = (float)$this->_order->getTaxAmount();
+        if($useOrderCurrency){
+            $orderRequest->currency = $this->_order->getOrderCurrencyCode();
+            $orderRequest->amount = $this->_order->getForcedAmount() ?: (float)$this->_order->getGrandTotal();
+            $orderRequest->shipping = (float)$this->_order->getShippingAmount();
+            $orderRequest->tax = (float)$this->_order->getTaxAmount();
+        }else{
+            $orderRequest->currency = $this->_order->getBaseCurrencyCode();
+            $orderRequest->amount = $this->_order->getForcedAmount() ?: (float)$this->_order->getBaseGrandTotal();
+            $orderRequest->shipping = (float)$this->_order->getBaseShippingAmount();
+            $orderRequest->tax = (float)$this->_order->getBaseTaxAmount();
+        }
+
         $orderRequest->cid = $this->_customerId;
         $orderRequest->ipaddr = $this->_order->getRemoteIp();
         $orderRequest->language = $this->_localeResolver->getLocale();
@@ -229,12 +251,22 @@ class Order extends CommonRequest
         $orderRequest->cancel_url = $this->_urlBuilder->getUrl('hipay/redirect/cancel', $redirectParams);
         $orderRequest->exception_url = $this->_urlBuilder->getUrl('hipay/redirect/exception', $redirectParams);
 
+        if ($this->_config->isSendingNotifyUrl()) {
+            $orderRequest->notify_url = $this->_urlBuilder->getUrl("hipay/notify/index");
+        }
+
         $orderRequest->paymentMethod = $this->_paymentMethod;
-        $orderRequest->customerBillingInfo = $this->_requestFactory->create('\HiPay\FullserviceMagento\Model\Request\Info\BillingInfo', ['params' => ['order' => $this->_order, 'config' => $this->_config]])->getRequestObject();
-        $orderRequest->customerShippingInfo = $this->_requestFactory->create('\HiPay\FullserviceMagento\Model\Request\Info\ShippingInfo', ['params' => ['order' => $this->_order, 'config' => $this->_config]])->getRequestObject();
+        $orderRequest->customerBillingInfo = $this->_requestFactory->create(
+            '\HiPay\FullserviceMagento\Model\Request\Info\BillingInfo',
+            ['params' => ['order' => $this->_order, 'config' => $this->_config]]
+        )->getRequestObject();
+        $orderRequest->customerShippingInfo = $this->_requestFactory->create(
+            '\HiPay\FullserviceMagento\Model\Request\Info\ShippingInfo',
+            ['params' => ['order' => $this->_order, 'config' => $this->_config]]
+        )->getRequestObject();
 
         // Extras informations
-        $this->processExtraInformations($orderRequest);
+        $this->processExtraInformations($orderRequest, $useOrderCurrency);
 
         return $orderRequest;
     }
@@ -244,7 +276,7 @@ class Order extends CommonRequest
      *
      * @param OrderRequest $order OrderRequest passed by reference
      */
-    private function processExtraInformations(OrderRequest &$orderRequest)
+    private function processExtraInformations(OrderRequest &$orderRequest, $useOrderCurrency = false)
     {
         // Check if fingerprint is enabled
         if ($this->_config->isFingerprintEnabled()) {
@@ -253,7 +285,7 @@ class Order extends CommonRequest
 
         // Check if sending cart is necessary ( If  conf enabled or if payment method product needs it )
         if ($this->_config->isNecessaryToSendCartItems($orderRequest->payment_product)) {
-            $orderRequest->basket = $this->processCartFromOrder($this->_operation);
+            $orderRequest->basket = $this->processCartFromOrder($this->_operation, $useOrderCurrency);
         }
 
         // Check if delivery method is required for the payment method